+++ tests/testsuite/registry.rs
@@ -14,7 +14,7 @@ use std::path::Path;
 
 fn cargo_http(p: &Project, s: &str) -> Execs {
     let mut e = p.cargo(s);
-    e.arg("-Zhttp-registry").masquerade_as_nightly_cargo();
+    e.arg("-Zsparse-registry").masquerade_as_nightly_cargo();
     e
 }
 
@@ -2643,13 +2643,13 @@ fn http_requires_z_flag() {
 
     p.cargo("build")
         .with_status(101)
-        .with_stderr_contains("  usage of HTTP-based registries requires `-Z http-registry`")
+        .with_stderr_contains("  usage of HTTP-based registries requires `-Z sparse-registry`")
         .run();
 }
 
 #[cargo_test]
 fn http_requires_trailing_slash() {
-    cargo_process("-Z http-registry install bar --index sparse+https://index.crates.io")
+    cargo_process("-Z sparse-registry install bar --index sparse+https://index.crates.io")
         .masquerade_as_nightly_cargo()
         .with_status(101)
         .with_stderr("[ERROR] registry url must end in a slash `/`: sparse+https://index.crates.io")